Be smarter about diffing when filters are used
When filtering commits, diff with the direct parent instead of the next
older commit that also matches the filters.
This shows fewer irrelevant changes for pickaxe or range queries.

Still use the chronologically next commit in the list if all commits for the file
are shown, this gives easier to follow diffs when some commits have multiple parents.